+++ tests/testthat/test-utils-metadata.R
@@ -17,7 +17,7 @@ test_that("feature_hits() works", {
   skip_on_cran()
   skip_if(geoserver_down)
   
-  expect_equal(vicmap_query("datavic:VMHYDRO_WATERCOURSE_DRAIN", wfs_version = "2.0.0") %>% 
-                 filter(HIERARCHY == "L", PFI == 8553127) %>%
+  expect_equal(vicmap_query("open-data-platform:hy_watercourse", wfs_version = "2.0.0") %>% 
+                 filter(hierarchy == "L", pfi == 8553127) %>%
                  feature_hits(), 1)
 })